aspects: generated ts file now support subclassing
I'm using a workaround to be able to export type nicely (__c & __i).
This approach is able to handle subclassing and extension of categories.
The only known issue yet is that subclassing categories only work as expected if there isn't a class in the hierarchy that doesn't define the category.